88 - Move store/constants to constants/store
