+++ src/store/comments.ts
@@ -1,6 +1,6 @@
 import {createSlice, PayloadAction} from '@reduxjs/toolkit';
 
-import {COMMENTS} from 'store/constants';
+import {COMMENTS} from 'constants/store';
 import {Comment, Comments} from 'types';
 
 const initialState: Comments = {};
