+++ src/store/exchangeOrders.ts
@@ -1,6 +1,6 @@
 import {createSlice, PayloadAction} from '@reduxjs/toolkit';
 
-import {EXCHANGE_ORDERS} from 'store/constants';
+import {EXCHANGE_ORDERS} from 'constants/store';
 import {ExchangeOrder, ExchangeOrders} from 'types';
 
 const initialState: ExchangeOrders = {};
