+++ src/store/invitationLimits.ts
@@ -1,6 +1,6 @@
 import {createSlice, PayloadAction} from '@reduxjs/toolkit';
 
-import {INVITATION_LIMITS} from 'store/constants';
+import {INVITATION_LIMITS} from 'constants/store';
 import {InvitationLimit, InvitationLimits} from 'types';
 
 const initialState: InvitationLimits = {};
