+++ src/store/invitations.ts
@@ -1,6 +1,6 @@
 import {createSlice, PayloadAction} from '@reduxjs/toolkit';
 
-import {INVITATIONS} from 'store/constants';
+import {INVITATIONS} from 'constants/store';
 import {Invitation, Invitations} from 'types';
 
 const initialState: Invitations = {};
