+++ src/store/trades.ts
@@ -1,6 +1,6 @@
 import {createSlice, PayloadAction} from '@reduxjs/toolkit';
 
-import {TRADES} from 'store/constants';
+import {TRADES} from 'constants/store';
 import {Trade, Trades} from 'types';
 
 const initialState: Trades = {};
