+++ src/store/wires.ts
@@ -1,6 +1,6 @@
 import {createSlice, PayloadAction} from '@reduxjs/toolkit';
 
-import {WIRES} from 'store/constants';
+import {WIRES} from 'constants/store';
 import {Wire, Wires} from 'types';
 
 const initialState: Wires = {};
